Move all annotation constants to shared location so any rules can access them
