--- lint/anvil/src/main/java/dev/whosnickdoglio/anvil/AnvilAnnotations.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright (C) 2024 Nicholas Doglio
- * SPDX-License-Identifier: MIT
- */
-package dev.whosnickdoglio.anvil
-
-internal const val CONTRIBUTES_TO = "com.squareup.anvil.annotations.ContributesTo"
-internal const val CONTRIBUTES_BINDING = "com.squareup.anvil.annotations.ContributesBinding"
-internal const val CONTRIBUTES_MULTI_BINDING =
-    "com.squareup.anvil.annotations.ContributesMultibinding"
-internal const val CONTRIBUTES_SUBCOMPONENT =
-    "com.squareup.anvil.annotations.ContributesSubcomponent"
-internal const val CONTRIBUTES_SUBCOMPONENT_FACTORY =
-    "com.squareup.anvil.annotations.ContributesSubcomponent.Factory"
-internal const val MERGE_COMPONENT = "com.squareup.anvil.annotations.MergeComponent"
-internal const val MERGE_SUBCOMPONENT = "com.squareup.anvil.annotations.MergeSubcomponent"
