+++ lint/anvil/src/main/java/dev/whosnickdoglio/anvil/detectors/MissingContributesToDetector.kt
@@ -15,8 +15,8 @@ import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.TextFormat
-import dev.whosnickdoglio.anvil.CONTRIBUTES_TO
-import dev.whosnickdoglio.lint.shared.MODULE
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_TO
+import dev.whosnickdoglio.lint.shared.dagger.MODULE
 import org.jetbrains.uast.UAnnotation
 import org.jetbrains.uast.UClass
 import org.jetbrains.uast.UElement
