+++ lint/anvil/src/main/java/dev/whosnickdoglio/anvil/detectors/NoAnvilInJavaDetector.kt
@@ -15,13 +15,13 @@ import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.TextFormat
-import dev.whosnickdoglio.anvil.CONTRIBUTES_BINDING
-import dev.whosnickdoglio.anvil.CONTRIBUTES_MULTI_BINDING
-import dev.whosnickdoglio.anvil.CONTRIBUTES_SUBCOMPONENT
-import dev.whosnickdoglio.anvil.CONTRIBUTES_SUBCOMPONENT_FACTORY
-import dev.whosnickdoglio.anvil.CONTRIBUTES_TO
-import dev.whosnickdoglio.anvil.MERGE_COMPONENT
-import dev.whosnickdoglio.anvil.MERGE_SUBCOMPONENT
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_MULTI_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_SUBCOMPONENT
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_SUBCOMPONENT_FACTORY
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_TO
+import dev.whosnickdoglio.lint.shared.anvil.MERGE_COMPONENT
+import dev.whosnickdoglio.lint.shared.anvil.MERGE_SUBCOMPONENT
 import org.jetbrains.uast.UAnnotation
 import org.jetbrains.uast.UElement
 import org.jetbrains.uast.kotlin.isKotlin
