+++ lint/anvil/src/test/java/dev/whosnickdoglio/anvil/detectors/ContributesBindingMustHaveSuperDetectorTest.kt
@@ -8,9 +8,9 @@ import com.android.tools.lint.checks.infrastructure.TestFiles
 import com.android.tools.lint.checks.infrastructure.TestLintTask
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import dev.whosnickdoglio.anvil.CONTRIBUTES_BINDING
-import dev.whosnickdoglio.anvil.CONTRIBUTES_MULTI_BINDING
-import dev.whosnickdoglio.anvil.CONTRIBUTES_TO
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_MULTI_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_TO
 import dev.whosnickdoglio.stubs.daggerAnnotations
 import dev.whosnickdoglio.stubs.javaxAnnotations
 import org.junit.Test
@@ -112,7 +112,7 @@ class ContributesBindingMustHaveSuperDetectorTest {
                     Fix for src/MyModule.kt line 4: Did you mean to use the `@ContributesTo` annotation?:
                     @@ -4 +4
                     - @${annotation.substringAfterLast(".")}
-                    + @${CONTRIBUTES_TO}
+                    + @$CONTRIBUTES_TO
                     +
                 """
                     .trimIndent(),
