+++ lint/anvil/src/test/java/dev/whosnickdoglio/anvil/detectors/MissingContributesBindingDetectorTest.kt
@@ -8,8 +8,8 @@ import com.android.tools.lint.checks.infrastructure.TestFiles
 import com.android.tools.lint.checks.infrastructure.TestLintTask
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import dev.whosnickdoglio.anvil.CONTRIBUTES_BINDING
-import dev.whosnickdoglio.anvil.CONTRIBUTES_MULTI_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_BINDING
+import dev.whosnickdoglio.lint.shared.anvil.CONTRIBUTES_MULTI_BINDING
 import dev.whosnickdoglio.stubs.javaxAnnotations
 import org.junit.Test
 import org.junit.runner.RunWith
