+++ lint/dagger/src/main/java/dev/whosnickdoglio/dagger/detectors/CorrectBindsUsageDetector.kt
@@ -15,7 +15,7 @@ import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.TextFormat
-import dev.whosnickdoglio.lint.shared.BINDS
+import dev.whosnickdoglio.lint.shared.dagger.BINDS
 import org.jetbrains.uast.UAnnotation
 import org.jetbrains.uast.UElement
 import org.jetbrains.uast.UMethod
