+++ lint/dagger/src/main/java/dev/whosnickdoglio/dagger/detectors/MultipleScopesDetector.kt
@@ -16,9 +16,9 @@ import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.TextFormat
-import dev.whosnickdoglio.lint.shared.BINDS
-import dev.whosnickdoglio.lint.shared.PROVIDES
-import dev.whosnickdoglio.lint.shared.SCOPE
+import dev.whosnickdoglio.lint.shared.dagger.BINDS
+import dev.whosnickdoglio.lint.shared.dagger.PROVIDES
+import dev.whosnickdoglio.lint.shared.dagger.SCOPE
 import org.jetbrains.uast.UAnnotated
 import org.jetbrains.uast.UClass
 import org.jetbrains.uast.UElement
