+++ lint/dagger/src/main/java/dev/whosnickdoglio/dagger/detectors/ScopedWithoutInjectAnnotationDetector.kt
@@ -15,8 +15,8 @@ import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
 import com.android.tools.lint.detector.api.TextFormat
-import dev.whosnickdoglio.lint.shared.INJECT
-import dev.whosnickdoglio.lint.shared.SCOPE
+import dev.whosnickdoglio.lint.shared.dagger.INJECT
+import dev.whosnickdoglio.lint.shared.dagger.SCOPE
 import org.jetbrains.uast.UAnnotated
 import org.jetbrains.uast.UClass
 import org.jetbrains.uast.UElement
