--- lint/hilt/src/main/java/dev/whosnickdoglio/hilt/HiltAnnotations.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2024 Nicholas Doglio
- * SPDX-License-Identifier: MIT
- */
-package dev.whosnickdoglio.hilt
-
-internal const val INSTALL_IN = "dagger.hilt.InstallIn"
-internal const val ENTRY_POINT = "dagger.hilt.EntryPoint"
-internal const val ANDROID_ENTRY_POINT = "dagger.hilt.android.AndroidEntryPoint"
-internal const val HILT_ANDROID_APP = "dagger.hilt.android.HiltAndroidApp"
-internal const val HILT_VIEW_MODEL = "dagger.hilt.android.lifecycle.HiltViewModel"
