+++ lint/hilt/src/test/java/dev/whosnickdoglio/hilt/detectors/MissingAndroidEntryPointDetectorTest.kt
@@ -8,7 +8,7 @@ import com.android.tools.lint.checks.infrastructure.TestFiles
 import com.android.tools.lint.checks.infrastructure.TestLintTask
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import dev.whosnickdoglio.hilt.ANDROID_ENTRY_POINT
+import dev.whosnickdoglio.lint.shared.hilt.ANDROID_ENTRY_POINT
 import dev.whosnickdoglio.stubs.javaxAnnotations
 import org.junit.Test
 import org.junit.runner.RunWith
