+++ lint/hilt/src/test/java/dev/whosnickdoglio/hilt/detectors/MissingHiltViewModelAnnotationDetectorTest.kt
@@ -6,7 +6,7 @@ package dev.whosnickdoglio.hilt.detectors
 
 import com.android.tools.lint.checks.infrastructure.TestFiles
 import com.android.tools.lint.checks.infrastructure.TestLintTask
-import dev.whosnickdoglio.hilt.HILT_VIEW_MODEL
+import dev.whosnickdoglio.lint.shared.hilt.HILT_VIEW_MODEL
 import dev.whosnickdoglio.stubs.javaxAnnotations
 import org.junit.Test
