+++ lint/shared/src/main/java/dev/whosnickdoglio/lint/shared/anvil/AnvilAnnotations.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright (C) 2024 Nicholas Doglio
+ * SPDX-License-Identifier: MIT
+ */
+package dev.whosnickdoglio.lint.shared.anvil
+
+public const val CONTRIBUTES_TO: String = "com.squareup.anvil.annotations.ContributesTo"
+public const val CONTRIBUTES_BINDING: String = "com.squareup.anvil.annotations.ContributesBinding"
+public const val CONTRIBUTES_MULTI_BINDING: String =
+    "com.squareup.anvil.annotations.ContributesMultibinding"
+public const val CONTRIBUTES_SUBCOMPONENT: String =
+    "com.squareup.anvil.annotations.ContributesSubcomponent"
+public const val CONTRIBUTES_SUBCOMPONENT_FACTORY: String =
+    "com.squareup.anvil.annotations.ContributesSubcomponent.Factory"
+public const val MERGE_COMPONENT: String = "com.squareup.anvil.annotations.MergeComponent"
+public const val MERGE_SUBCOMPONENT: String = "com.squareup.anvil.annotations.MergeSubcomponent"
