+++ lint/shared/src/main/java/dev/whosnickdoglio/lint/shared/hilt/HiltAnnotations.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2024 Nicholas Doglio
+ * SPDX-License-Identifier: MIT
+ */
+package dev.whosnickdoglio.lint.shared.hilt
+
+public const val INSTALL_IN: String = "dagger.hilt.InstallIn"
+public const val ENTRY_POINT: String = "dagger.hilt.EntryPoint"
+public const val ANDROID_ENTRY_POINT: String = "dagger.hilt.android.AndroidEntryPoint"
+public const val HILT_ANDROID_APP: String = "dagger.hilt.android.HiltAndroidApp"
+public const val HILT_VIEW_MODEL: String = "dagger.hilt.android.lifecycle.HiltViewModel"
